# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Platform API
  description: |
    Agent Platform Product Experience API - A multi-agent platform for task orchestration 
    and execution with LLM providers, background task processing, workflows, triggers, and connectors.
    
    All endpoints use a standard API response envelope:
    - Success: `{ ok: true, data: ..., requestId: "..." }`
    - Error: `{ ok: false, error: { code: "...", message: "..." }, requestId: "..." }`
  version: "0.7.0-rc.1"
  contact:
    name: Agent Platform Team
servers:
  - url: http://localhost:3003
    description: Development
  - url: http://localhost:3103
    description: E2E Testing
...
paths:
...
  /api/v1/triggers/schedules:
    get:
      tags: [Triggers]
      summary: List schedule triggers
      description: Returns all schedule triggers for the current user
      operationId: listScheduleTriggers
      security:
        - sessionAuth: []
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Triggers]
      summary: Create schedule trigger
      description: Creates a new schedule trigger
      operationId: createScheduleTrigger
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleTriggerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/triggers/schedules/{scheduleId}:
    get:
      tags: [Triggers]
      summary: Get schedule trigger
      description: Returns a specific schedule trigger
      operationId: getScheduleTrigger
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/scheduleId"
      responses:
        "200":
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleTriggerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Triggers]
      summary: Update schedule trigger
      description: Updates a schedule trigger status
      operationId: updateScheduleTrigger
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/scheduleId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTriggerStatusRequest"
      responses:
        "200":
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleTriggerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Triggers]
      summary: Delete schedule trigger
      description: Deletes a schedule trigger
      operationId: deleteScheduleTrigger
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/scheduleId"
      responses:
        "204":
          description: Schedule deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/triggers/webhooks:
    get:
      tags: [Triggers]
      summary: List webhook triggers
      description: Returns all webhook triggers for the current user
      operationId: listWebhookTriggers
      security:
        - sessionAuth: []
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Triggers]
      summary: Create webhook trigger
      description: Creates a new webhook trigger
      operationId: createWebhookTrigger
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTriggerCreatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/triggers/webhooks/{webhookId}:
    get:
      tags: [Triggers]
      summary: Get webhook trigger
      description: Returns a specific webhook trigger
      operationId: getWebhookTrigger
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/webhookId"
      responses:
        "200":
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTriggerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Triggers]
      summary: Update webhook trigger
      description: Updates a webhook trigger status
      operationId: updateWebhookTrigger
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/webhookId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTriggerStatusRequest"
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTriggerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Triggers]
      summary: Delete webhook trigger
      description: Deletes a webhook trigger
      operationId: deleteWebhookTrigger
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/webhookId"
      responses:
        "204":
          description: Webhook deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: agent-platform-session
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API token for programmatic access (set API_AUTH_TOKEN env var to enable)
  parameters:
...
    scheduleId:
      name: scheduleId
      in: path
      required: true
      schema:
        type: string
      description: Schedule trigger ID
    webhookId:
      name: webhookId
      in: path
      required: true
      schema:
        type: string
      description: Webhook trigger ID
...
  schemas:
...
    ApiEnvelope:
      type: object
      required: [ok, requestId]
      properties:
        ok:
          type: boolean
          description: true for success, false for error
        requestId:
          type: string
          description: Traceable request identifier
    ApiErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          description: Additional error details
...
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [error]
          properties:
            error:
              $ref: "#/components/schemas/ApiErrorDetail"

    # ---- Auth & Setup ----
...
    ScheduleTriggerResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [scheduleId, name, schedulePattern, status, runCount, createdAt, updatedAt]
              properties:
                scheduleId:
                  type: string
                name:
                  type: string
                schedulePattern:
                  type: string
                status:
                  type: string
                lastRunAt:
                  type: string
                  format: date-time
                nextRunAt:
                  type: string
                  format: date-time
                runCount:
                  type: integer
                maxRuns:
                  type: integer
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time
    ScheduleListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ScheduleTriggerResponse/properties/data"
    CreateScheduleRequest:
      type: object
      required: [name, schedulePattern]
      properties:
        name:
          type: string
        schedulePattern:
          type: string
        maxRuns:
          type: integer
    WebhookTriggerResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [webhookId, name, status, secretLast4, createdAt, updatedAt]
              properties:
                webhookId:
                  type: string
                name:
                  type: string
                status:
                  type: string
                secretLast4:
                  type: string
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time
    WebhookTriggerCreatedResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [webhookId, name, status, secretLast4, secret, createdAt, updatedAt]
              properties:
                webhookId:
                  type: string
                name:
                  type: string
                status:
                  type: string
                secretLast4:
                  type: string
                secret:
                  type: string
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time
    WebhookListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/WebhookTriggerResponse/properties/data"
    CreateWebhookRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateTriggerStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, paused]
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            requestId: "abc12345"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            requestId: "abc12345"
...